pyproject.toml
[tool.poetry]
name = "crypto-rank-tracker"
version = "0.2.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
aiohttp = "^3.12.13"
google-cloud-storage = "^3.1.1"
aiofiles = "^24.1.0"
tenacity = "^9.1.2"
pydantic = "^2.11.7"
numpy = ">=2.0,<2.3"
tqdm = "^4.67.1"
functions-framework = "^3.9.2"
aiolimiter = "^1.2.1"
pandas = "^2.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
